pyproject.toml
[project]
name = "rentcar"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "django-ckeditor-5>=0.2.18",
    "django-filter>=25.2",
    "django-jazzmin>=3.0.1",
    "django-storages[s3]>=1.14.6",
    "django[argon2]>=5.2.7",
    "djangorestframework-simplejwt>=5.5.1",
    "dotenv>=0.9.9",
    "drf-spectacular>=0.28.0",
    "flake8>=6.1.0",
    "isort>=6.1.0",
    "minio-storage>=0.1.5",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "redis>=6.4.0",
    "ruff>=0.14.0",
]
